add valiadtion details
